NicknameInput 내부 함수 추가
